+++ src/Presentation/ViewModel/School/SchoolImagesResponseViewModel.cs
@@ -10,7 +10,7 @@ namespace GamaEdtech.Presentation.ViewModel.School
         public DateTimeOffset CreationDate { get; set; }
         public string? FileId { get; set; }
         public FileType? FileType { get; set; }
-        public int SchoolId { get; set; }
+        public long SchoolId { get; set; }
         public string? SchoolName { get; set; }
     }
 }
